feat: pagamentos para inserir no banco de dados

+++ database/migrations/2025_06_13_014458_create_mod_vendas_item_pedido_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('id_pedido');
             $table->unsignedBigInteger('id_produto');
             $table->integer('quantidade');
-            $table->decimal('valor_unitario', 10, 2);
+            $table->integer('valor_unitario');
             $table->timestamps();
         });
     }
